+++ include/matx/transforms/matmul.h
@@ -1082,10 +1082,12 @@ private:
 struct MatMulParamsKeyHash {
   std::size_t operator()(const MatMulParams_t &k) const noexcept
   {
-    return std::hash<uint64_t>()(k.m) + std::hash<uint64_t>()(k.n) +
-           std::hash<uint64_t>()(k.k) + std::hash<uint64_t>()(k.batch) +
-           std::hash<uint64_t>()(k.prov) +
-           std::hash<uint64_t>()((size_t)k.stream);
+    return  std::hash<uint64_t>()(static_cast<uint64_t>(k.m)) + 
+            std::hash<uint64_t>()(static_cast<uint64_t>(k.n)) +
+            std::hash<uint64_t>()(static_cast<uint64_t>(k.k)) + 
+            std::hash<uint64_t>()(static_cast<uint64_t>(k.batch)) +
+            std::hash<uint64_t>()(static_cast<uint64_t>(k.prov)) +
+            std::hash<uint64_t>()((uint64_t)k.stream);
   }
 };
