+++ include/matx/transforms/solver.h
@@ -298,7 +298,8 @@ private:
 struct DnCholParamsKeyHash {
   std::size_t operator()(const DnCholParams_t &k) const noexcept
   {
-    return (std::hash<uint64_t>()(k.n)) + (std::hash<uint64_t>()(k.batch_size));
+    return  (std::hash<uint64_t>()(static_cast<uint64_t>(k.n))) + 
+            (std::hash<uint64_t>()(static_cast<uint64_t>(k.batch_size)));
   }
 };
 
@@ -469,8 +470,9 @@ private:
 struct DnLUParamsKeyHash {
   std::size_t operator()(const DnLUParams_t &k) const noexcept
   {
-    return (std::hash<uint64_t>()(k.m)) + (std::hash<uint64_t>()(k.n)) +
-           (std::hash<uint64_t>()(k.batch_size));
+    return  (std::hash<uint64_t>()(static_cast<uint64_t>(k.m))) + 
+            (std::hash<uint64_t>()(static_cast<uint64_t>(k.n))) +
+            (std::hash<uint64_t>()(static_cast<uint64_t>(k.batch_size)));
   }
 };
 
@@ -649,8 +651,9 @@ private:
 struct DnQRParamsKeyHash {
   std::size_t operator()(const DnQRParams_t &k) const noexcept
   {
-    return (std::hash<uint64_t>()(k.m)) + (std::hash<uint64_t>()(k.n)) +
-           (std::hash<uint64_t>()(k.batch_size));
+    return (std::hash<uint64_t>()(static_cast<uint64_t>(k.m))) + 
+           (std::hash<uint64_t>()(static_cast<uint64_t>(k.n))) +
+           (std::hash<uint64_t>()(static_cast<uint64_t>(k.batch_size)));
   }
 };
 
@@ -868,8 +871,9 @@ private:
 struct DnSVDParamsKeyHash {
   std::size_t operator()(const DnSVDParams_t &k) const noexcept
   {
-    return (std::hash<uint64_t>()(k.m)) + (std::hash<uint64_t>()(k.n)) +
-           (std::hash<uint64_t>()(k.batch_size));
+    return (std::hash<uint64_t>()(static_cast<uint64_t>(k.m))) + 
+           (std::hash<uint64_t>()(static_cast<uint64_t>(k.n))) +
+           (std::hash<uint64_t>()(static_cast<uint64_t>(k.batch_size)));
   }
 };
 
@@ -1063,7 +1067,8 @@ private:
 struct DnEigParamsKeyHash {
   std::size_t operator()(const DnEigParams_t &k) const noexcept
   {
-    return (std::hash<uint64_t>()(k.m)) + (std::hash<uint64_t>()(k.batch_size));
+    return (std::hash<uint64_t>()(static_cast<uint64_t>(k.m))) + 
+           (std::hash<uint64_t>()(static_cast<uint64_t>(k.batch_size)));
   }
 };
